+++ src/app/pages/tabs/analysis/analysis-list/analysis-list.page.spec.ts
@@ -1,0 +1,24 @@
+import { ComponentFixture, TestBed, waitForAsync } from "@angular/core/testing";
+import { IonicModule } from "@ionic/angular";
+
+import { AnalysisListPage } from "./analysis-list.page";
+
+describe("AnalysisListPage", () => {
+  let component: AnalysisListPage;
+  let fixture: ComponentFixture<AnalysisListPage>;
+
+  beforeEach(waitForAsync(() => {
+    TestBed.configureTestingModule({
+      declarations: [AnalysisListPage],
+      imports: [IonicModule.forRoot()],
+    }).compileComponents();
+
+    fixture = TestBed.createComponent(AnalysisListPage);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  }));
+
+  it("should create", () => {
+    expect(component).toBeTruthy();
+  });
+});