+++ src/app/pages/tabs/reports/reports-routing.module.ts
@@ -1,13 +1,18 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
+import { NgModule } from "@angular/core";
+import { Routes, RouterModule } from "@angular/router";
 
-import { ReportsPage } from './reports.page';
+import { ReportsPage } from "./reports.page";
+import { ReportDetailsPage } from "./report-details/report-details.page";
 
 const routes: Routes = [
   {
-    path: '',
-    component: ReportsPage
-  }
+    path: "",
+    component: ReportsPage,
+  },
+  {
+    path: "report-details",
+    component: ReportDetailsPage,
+  },
 ];
 
 @NgModule({